ref: вынести обработчик ошибок в отдельный модуль
